Rosie branch

+++ GameMechanics/src/main/java/org/game/entity/Quiz.java
@@ -17,6 +17,6 @@ public class Quiz {
     private int quizId;
     private String quizName;
     private boolean isExam;
-    private int questionNumber; // eg. a normal quiz may be 10 questions, an exam may be 20
+    private byte questionNumber; // eg. a normal quiz may be 10 questions, an exam may be 20
     private int xpWorth; // The xp the player will be rewarded with
 }
